* commit 'ff4af1a11cfec1a7fb32f76d7e1446ff6860be53 [formerly 38598e3071b93fc20b0bec2c4f4df8ce6a8358dc]': (200 commits)
  [DF-185] Code format and script format
  [DF-12604] file moved
  [DF-185] Fixed warnings
  [DF-185] Warnings fix
  DF-12294 code style fixes
  DF-12294 code style fixes
  [DF-185] Fix Warnings
  [DF-185] Fix Warnings
  DF-12591 Format code
  DF-12591 Add function to check element color
  [DF-185] Disable autotests for every project
  DF-12294 compilation fix
  DF-12169 fix compiler warnings in AutotestingSystem_wrap.cxx
  DF-12169 testbed: in FullscreenTest now you can switch to fullscreen and pinning and back using keyboard
  DF-12169 introduce WINDOW_CANCEL_INPUT to cancel inputs in cases other than focus changing
  DF-12169 win32: disable 'move' and 'size' commands from window's system menu
  DF-12574_subsystem_console fix subsystem
  [DF-185] fixed cmake message type
  [DF-12558] Now autotesting screenshots texture depends on window size
  DF-12169 win32: fix switching to pinning when mouse button is pressed on win7
  ...


Former-commit-id: 5b9fa361207f037984f837c0130ad6f76ab8a0cd

+++ Modules/TArc/Sources/TArc/WindowSubSystem/Private/UIManager.cpp
@@ -309,11 +309,11 @@ void AddStatusbarPoint(const QUrl& url, QAction* action, MainWindowInfo& windowI
     QStatusBar* statusBar = windowInfo.window->statusBar();
     if (isPermanent)
     {
-        statusBar->insertPermanentWidget(positionIndex, actionWidget, stretchFactor);
+        statusBar->insertPermanentWidget(static_cast<int>(positionIndex), actionWidget, stretchFactor);
     }
     else
     {
-        statusBar->insertWidget(positionIndex, actionWidget, stretchFactor);
+        statusBar->insertWidget(static_cast<int>(positionIndex), actionWidget, stretchFactor);
     }
 }
